So, how about this?
The idea is that if get_statements returns no statements, no further
processing is necessary, and we should return an empty iterator as
quickly as possibly. Now, I don't know if $_iter->finished would
actually do this, but it seems intuitively right to me (if there are
no statements, we're finished although we haven't called ->next). And
I'm not sure the iterator would be empty like this, but it passes my
tests...