API inconsistent between RDF::Trine::Store::Memory and RDF::Trine::Store::SPARQL #149
As per issue #149 on Github, RDF::Trine::Store::SPARQL doesn't implement
the remove_statements() function interface in accordance with the documentation
or all the other RDF::Trine::Store::* modules. It takes ($statement,$context)
parameters instead of ($subj,$pred,$ojb,$context) parameters.

This patch keeps this inconsistent behaviour so as to be backwards
compatible, but also provides support for the interface referenced
in the documentation and implemented by other storage backends.

This patch also includes unit tests to check both sets of parameters
with and without context.